+++ tests/test_linter.py
@@ -7,7 +7,7 @@ run::
 
    # for a single test
 
-   pytest tests/test_linter.py::test_H019 --cov=src/djlint --cov-branch \
+   pytest tests/test_linter.py::test_H021 --cov=src/djlint --cov-branch \
          --cov-report xml:coverage.xml --cov-report term-missing
 
 """
@@ -276,6 +276,13 @@ def test_H020(runner: CliRunner, tmp_file: TextIO) -> None:
     result = runner.invoke(djlint, [tmp_file.name])
     assert result.exit_code == 1
     assert "H020 1:" in result.output
+
+
+def test_H021(runner: CliRunner, tmp_file: TextIO) -> None:
+    write_to_file(tmp_file.name, b'<div style="asdf"></div>')
+    result = runner.invoke(djlint, [tmp_file.name])
+    assert result.exit_code == 1
+    assert "H021 1:" in result.output
 
 
 def test_rules_not_matched_in_ignored_block(